backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
langchain>=0.1.13
langchain-community>=0.0.29
langchain-core>=0.1.41
langchain-groq>=0.1.1
wikipedia==1.4.0
pydantic>=2.9.0
langchain-google-genai==1.0.1
langchain-anthropic==0.1.6
langchain-openai==0.1.1
python-dotenv==1.0.1
